==> 2_Data_wrangling.Rmd <==
---
title: "Data Wrangling"
author: "Cheryn Aouaj, Lisa Bebell, Sun Kim, Manasi Mohan"
date: "12/10/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

**_First we will import the raw data, clean them, and build a few summary variables that are of interest for our analyses. Then we will take monthly averages which will allow us to appropriately merge the data._**
```{r, message=FALSE}
library(readr)
library(dplyr)
library(lubridate)
library(ggplot2)
```

##### Sea water temperature data (Moore Reef)
```{r}
# function to clean and wrangle AIMS Moore Reef data to deal with column names and column placement
clean_MooreReef_data <- function(df) {
    df[,2:ncol(df)] <- df[,1:ncol(df)]
    df[,1] <- rownames(df)
    df <- df %>% select(date, colnames(df)[ncol(df)])
    return(df)
}
```

```{r}
# read data
setwd("raw_data")
moore_reef_water_temp_2.0m <- read.csv("AIMS_MooreReef_WaterTemperature_22Oct1997to16Feb2020_2.0m.csv", skip=108, sep= ",", row.names = NULL)
moore_reef_water_temp_9.0m <- read.csv("AIMS_MooreReef_WaterTemperature_22Oct1997to17Dec2017_9.0m.csv", skip=94, sep= ",", row.names = NULL)

# run through function defined above to clean and wrangle data
moore_reef_water_temp_2.0m <- clean_MooreReef_data(moore_reef_water_temp_2.0m)
moore_reef_water_temp_9.0m <- clean_MooreReef_data(moore_reef_water_temp_9.0m)

moore_reef_water_temp_2.0m <- moore_reef_water_temp_2.0m %>%
    filter(Water.Temperature..2.0m.MORFL1.Reef.Flat.Site.1_LEVEL2_value_AVG != "Not available")

moore_reef_water_temp_9.0m <- moore_reef_water_temp_9.0m %>%
    filter(Water.Temperature..9.0m.MORSL1.Reef.Slope.Site.1_LEVEL2_value_AVG != "Not available")
```

```{r}
# merge AIMS Moore reef temperature data
aims_temp_data <- Reduce(function(x,y) merge(x,y, by="date"), list(moore_reef_water_temp_2.0m,
                                                              moore_reef_water_temp_9.0m))
                                                            
# convert water temp data from string to numeric type
aims_temp_data$Water.Temperature..2.0m.MORFL1.Reef.Flat.Site.1_LEVEL2_value_AVG <- 
    as.numeric(as.character(aims_temp_data$Water.Temperature..2.0m.MORFL1.Reef.Flat.Site.1_LEVEL2_value_AVG))

aims_temp_data$Water.Temperature..9.0m.MORSL1.Reef.Slope.Site.1_LEVEL2_value_AVG <- 
    as.numeric(as.character(aims_temp_data$Water.Temperature..9.0m.MORSL1.Reef.Slope.Site.1_LEVEL2_value_AVG))

# convert to date column into date type
aims_temp_data$date <- as.Date(aims_temp_data$date)

colnames(aims_temp_data) <- c("date", "avg_water_temp_2.0m_flat_site", "avg_water_temp_9.0m_slope_site")
```

```{r}
# save clean temperature file
setwd("cleaned_data")
write.csv(aims_temp_data, "aims_temperatures.csv", row.names = FALSE)

# clear environment
rm(list = ls())
```

##### Hard coral cover data
```{r}
# read data
setwd("raw_data")
coral_cover <- read.csv("trendgbr-coral-cover-with-ci.csv")

# convert date-decimal in coral cover to YYYY-MM-DD format
coral_cover$Date <- as.Date(format(date_decimal(coral_cover$Date), "%Y-%m-%d"))

# rename "Date" to "date"
names(coral_cover)[names(coral_cover) == "Date"] <- "date"
colnames(coral_cover) <- c("date", "mean_live_coral_cover_percent", "lower_conf_int", "upper_conf_int", "conf_int_span")
```

```{r}
# save clean coral cover file
setwd("cleaned_data")
write.csv(coral_cover, "coral_cover.csv", row.names = FALSE)

# clear environment
rm(list = ls())
```

##### Visual fish census data
```{r}
# read data
setwd("raw_data")
fish_census <- read.csv("Fish census 1992-2015.csv", sep="\t", header=T)

# select relevant columns
fish_census <- fish_census %>% select(gbifID, class, family, genus, species, verbatimScientificName, decimalLatitude, decimalLongitude, dateIdentified)

# reformat date
fish_census$dateIdentified <- ymd_hms(fish_census$dateIdentified)

# create new df with fish counts by date
fish_species_counts <- fish_census %>%
    arrange(dateIdentified) %>%
    group_by(dateIdentified) %>%
    summarise(num_of_species=n_distinct(species))

# rename "dateIdentified" to "date"
names(fish_species_counts)[names(fish_species_counts) == "dateIdentified"] <- "date"

# reformat date
fish_species_counts$date <- as.Date(fish_species_counts$date)
```

```{r}
# save clean fish counts file
setwd("cleaned_data")
write.csv(fish_species_counts, "fish_species_counts.csv", row.names = FALSE)

# clear environment
rm(list = ls())
```

##### Seagrass surveys
We decided not to use seagrass species belonging to the *Halophila* genus because they are so widespread in tropical waters. Since these species were present in nearly all the survey sites in our data, they would add little information to our models, so we determined that it did not make sense to include them.

```{r}
# read data
setwd("raw_data")
seagrass_data <- read.csv("GBR_NESP-TWQ-3.2.1-5.4_JCU_Seagrass_1984-2018_Site-surveys.csv") %>%
    # filter out for rows where we have information
    filter(SEDIMENT != "Not recorded" & PRESENCE_A == "Present") %>%

    # delete columns we are not going to use
    select(-FID, -MONTH, -YEAR, 
           -SURVEY_MET, -SURVEY_NAM, 
           # remove seagrass belonging to halophila genus
           -H_CAPRICOR, -H_TRICOSTA, -H_OVALIS, -H_UNINERVI, -H_DECIPIEN, -H_SPINULOS) 
```

```{r}
# examine how many presence/absence we see for each sea grass species
table(seagrass_data$C_ROTUNDAT)
table(seagrass_data$C_SERRULAT)
table(seagrass_data$E_ACOROIDE)
table(seagrass_data$S_ISOETIFO)
table(seagrass_data$T_CILIATUM)
table(seagrass_data$T_HEMPRICH)
table(seagrass_data$Z_CAPRICOR)
```

We see that there is actually no data for presence of `T_CILIATUM`. In addition, there are only 59 observations for `E_ACOROIDE` and 187 observations for `C_ROTUNDAT` that are actually useful in classifying these two species. Because we have a rather large data set, we want 200 or more observations for our classification model. So, we remove these columns from our classification problem due to insufficient data. After deleting some columns, there may be some observations where all the remaining species columns have "No" values for absence. So, we keep only rows where presence of at least one species was observed.
```{r}
# remove columns that provide little to no information (<200 observations)
seagrass_data <- seagrass_data %>% select(-C_ROTUNDAT, -T_CILIATUM, -E_ACOROIDE)

# remove rows with no seagrass observations
seagrass_data <- seagrass_data %>%
    filter(C_SERRULAT=="Yes" | 
           S_ISOETIFO=="Yes" |
           T_HEMPRICH=="Yes" |
           Z_CAPRICOR=="Yes")
```

```{r}
# build function to count number of seagrass species
count_species_present <- function(C_SERRULAT, S_ISOETIFO, T_HEMPRICH, Z_CAPRICOR) {
    count = 0
    
    if (C_SERRULAT=="Yes") {
        count <- count + 1
    }
    if (S_ISOETIFO=="Yes") {
        count <- count + 1
    }
    if (T_HEMPRICH=="Yes") {
        count <- count + 1
    }
    if (Z_CAPRICOR=="Yes") {
        count <- count + 1
    }
    
    return(count)
}

# count species
seagrass_data$num_species_present <- mapply(count_species_present, 
                                            seagrass_data$C_SERRULAT,
                                            seagrass_data$S_ISOETIFO,
                                            seagrass_data$T_HEMPRICH,
                                            seagrass_data$Z_CAPRICOR)

# how many survey sites had more than 1 species discovered?
table(seagrass_data$num_species_present)

nrow(seagrass_data)
```

We see that only 3.2% (414/12795) of total survey sites recorded more than 1 species observed. For ease of building a classification model, we will remove these rows. We do not want a situation where our model "can't decide" in classifying our observations. Because we removed only about 3% of over 12,500 observations, we still preserve statistical power.
```{r}
seagrass_data <- seagrass_data %>% filter(num_species_present < 2)

table(seagrass_data$C_SERRULAT)
table(seagrass_data$S_ISOETIFO)
table(seagrass_data$T_HEMPRICH)
table(seagrass_data$Z_CAPRICOR)
```

So, after some data wrangling and exploratory analysis, we will build a model to classify 4 species of seagrass: *Cymodocea serrulata*, *Syringodium isoetifolium*, *Thalassia hemprichii*, and *Zostera muelleri* (subspecies *capricorni*). Now, let's build a `SPECIES` column that collects all the presence in a single variable.
```{r}
# function to make a species column
get_species_type <- function(C_SERRULAT, S_ISOETIFO, T_HEMPRICH, Z_CAPRICOR) {
    if (C_SERRULAT=="Yes") {
        return("C_SERRULAT")
    } else if (S_ISOETIFO=="Yes") {
        return("S_ISOETIFO")
    } else if (T_HEMPRICH =="Yes") {
        return("T_HEMPRICH")
    } else if (Z_CAPRICOR =="Yes") {
        return("Z_CAPIRCOR")
    }
}

# build species column to classify species of each observation based on presence/absence
seagrass_data$SPECIES <- mapply(get_species_type, 
                                seagrass_data$C_SERRULAT, 
                                seagrass_data$S_ISOETIFO,
                                seagrass_data$T_HEMPRICH,
                                seagrass_data$Z_CAPRICOR)

table(seagrass_data$SPECIES)
```

```{r}
# convert SPECIES to unordered factor
seagrass_data$SPECIES <- factor(seagrass_data$SPECIES, ordered=FALSE)

# rename misspelled column in original data set
names(seagrass_data)[names(seagrass_data) == "LATITUTDE"] <- "LATITUDE"

# only select columns relevant for our ML algorithm
seagrass_data <- seagrass_data %>% 
    select(SPECIES, LATITUDE, LONGITUDE, DEPTH, SEDIMENT, TIDAL)
```

```{r}
# save cleaned seagrass data
setwd("cleaned_data")
write.csv(seagrass_data, "seagrass_classification_data.csv", row.names = FALSE)

# clear environment
rm(list = ls())
```


**_Now that each individual dataset has been cleaned, we need to merge them for ease of analysis. Since the seagrass dataset only includes month and year information (no specific dates), we chose to build our aggregate dataset at the monthly level by taking monthly averages of each variable._**

##### Sea water temperature data (Moore Reef)
```{r, message=FALSE, warning=FALSE}
setwd("raw_data")
temp2 <- read_csv("AIMS_MooreReef_WaterTemperature_22Oct1997to16Feb2020_2.0m.csv", skip=108)
temp9 <- read_csv("AIMS_MooreReef_WaterTemperature_22Oct1997to17Dec2017_9.0m.csv", skip=94)

temp2 <- temp2[,c(2,5)]
temp9 <- temp9[,c(2,5)]

temp2$date <- ymd(temp2$date)
temp9$date <- ymd(temp9$date)

temp <- merge(temp2, temp9, by = "date", all = TRUE)
temp <- temp %>%
  rename(temp_2m_flat = `Water Temperature @2.0m MORFL1 Reef Flat Site 1_LEVEL2_value_AVG`,
         temp_9m_slope = `Water Temperature @9.0m MORSL1 Reef Slope Site 1_LEVEL2_value_AVG`)
rm(temp2, temp9)

temp <- temp[!(is.na(temp$temp_2m_flat) & is.na(temp$temp_9m_slope)), ]

temp <- temp %>%
  mutate(year = year(date),
         month = month(date))

temp_agg2 <- aggregate(temp_2m_flat ~ year + month, data = temp, FUN = mean)
temp_agg9 <- aggregate(temp_9m_slope ~ year + month, data = temp, FUN = mean)

temp_agg2$monthyear <- my(paste(temp_agg2$month, temp_agg2$year, sep = " "))
temp_agg9$monthyear <- my(paste(temp_agg9$month, temp_agg9$year, sep = " "))

temp_agg <- merge(temp_agg2, temp_agg9, by = "monthyear", all = TRUE)
temp_agg <- temp_agg[, c('monthyear', 'temp_2m_flat', 'temp_9m_slope')]
rm(temp_agg2, temp_agg9)

temp_agg <- temp_agg %>%
  mutate(water_temp = rowMeans(temp_agg[,c('temp_2m_flat', 'temp_9m_slope')], na.rm = TRUE))
```

##### Hard coral cover data
```{r}
# read data
setwd("raw_data")
coral <- read.csv("trendgbr-coral-cover-with-ci.csv")

# reformat date
coral$Date <- format(date_decimal(coral$Date), "%Y-%m-%d")
coral$Date <- ymd(coral$Date)

# rename date and coral columns
coral <- coral %>%
  rename(date = Date,
         percent_coral_cover = `Mean.live.coral.cover..percent.`)

# add year and month columns
coral <- coral %>%
  mutate(year = year(date),
         month = month(date))

# create monthly aggregated set
coral_agg <- aggregate(percent_coral_cover ~ year + month, data = coral, FUN = mean)

# add monthyear column
coral_agg$monthyear <- my(paste(coral_agg$month, coral_agg$year, sep = " "))

# add season column
coral_agg <- coral_agg %>%
  mutate(Season = ifelse(month(monthyear) %in% c(10,11,12,1,2,3), "Wet/Hot (Oct-Mar)", "Dry/Cool (Apr-Sept)"))
```

##### Visual fish census data
```{r, message=FALSE, warning=FALSE}
# read data
setwd("raw_data")
fish <- read_delim("Fish census 1992-2015.csv", "\t", escape_double = FALSE, trim_ws = TRUE)

# select relevant columns
fish <- fish %>%
  select(gbifID, class, family, genus, species, verbatimScientificName, decimalLatitude, decimalLongitude, dateIdentified)

# reformat date
fish$dateIdentified <- ymd(fish$dateIdentified)

# grouping by date, count unique fish species
fish <- fish %>%
    arrange(dateIdentified) %>%
    group_by(dateIdentified) %>%
    summarise(num_fish=n_distinct(species))

# rename "dateIdentified" to "date"
fish <- fish %>%
  rename(date = dateIdentified)

# add year and month columns
fish <- fish %>%
  mutate(year = year(date),
         month = month(date))

# create monthly aggregated set
fish_agg <- aggregate(num_fish ~ year + month, data = fish, FUN = mean)

# add monthyear column
fish_agg$monthyear <- my(paste(fish_agg$month, fish_agg$year, sep = " "))
```

##### Seagrass surveys
```{r, message=FALSE, warning=FALSE}
setwd("raw_data")
seagrass <- read_csv("GBR_NESP-TWQ-3.2.1-5.4_JCU_Seagrass_1984-2018_Site-surveys.csv")

# convert to boolean
seagrass$PRESENCE_A <- ifelse(seagrass$PRESENCE_A=="Present", 1, 0)
seagrass$C_ROTUNDAT <- ifelse(seagrass$C_ROTUNDAT=="Yes", 1, 0)
seagrass$C_SERRULAT <- ifelse(seagrass$C_SERRULAT=="Yes", 1, 0)
seagrass$E_ACOROIDE <- ifelse(seagrass$E_ACOROIDE=="Yes", 1, 0)
seagrass$H_CAPRICOR <- ifelse(seagrass$H_CAPRICOR=="Yes", 1, 0)
seagrass$H_DECIPIEN <- ifelse(seagrass$H_DECIPIEN=="Yes", 1, 0)
seagrass$H_OVALIS <- ifelse(seagrass$H_OVALIS=="Yes", 1, 0)
seagrass$H_SPINULOS <- ifelse(seagrass$H_SPINULOS=="Yes", 1, 0)
seagrass$H_TRICOSTA <- ifelse(seagrass$H_TRICOSTA=="Yes", 1, 0)
seagrass$H_UNINERVI <- ifelse(seagrass$H_UNINERVI=="Yes", 1, 0)
seagrass$S_ISOETIFO <- ifelse(seagrass$S_ISOETIFO=="Yes", 1, 0)
seagrass$T_CILIATUM <- ifelse(seagrass$T_CILIATUM=="Yes", 1, 0)
seagrass$T_HEMPRICH <- ifelse(seagrass$T_HEMPRICH=="Yes", 1, 0)
seagrass$Z_CAPRICOR <- ifelse(seagrass$Z_CAPRICOR=="Yes", 1, 0)

# add count variable
seagrass <- seagrass %>%
  group_by(FID) %>%
  mutate(num_seagrass = sum(C_ROTUNDAT, C_SERRULAT, E_ACOROIDE, H_CAPRICOR, H_DECIPIEN,
                            H_OVALIS, H_SPINULOS, H_TRICOSTA, H_UNINERVI, S_ISOETIFO,
                            T_CILIATUM, T_HEMPRICH, Z_CAPRICOR))

unique(seagrass$num_seagrass)

# add monthyear column
seagrass$monthyear <- my(paste(seagrass$MONTH, seagrass$YEAR, sep = " "))

# create monthly aggregated set for number
seagrass_agg <- aggregate(num_seagrass ~ monthyear, data = seagrass, FUN = mean)

# create monthly aggregated set for boolean
seagrass_presence <- aggregate(PRESENCE_A ~ monthyear, data = seagrass, FUN = max)
```

##### Build and save aggregate dataframe
```{r}
# series of joins
features_agg <- left_join(coral_agg[, c('monthyear', 'Season', 'percent_coral_cover')], fish_agg[, c('monthyear', 'num_fish')], by = "monthyear")
features_agg <- left_join(features_agg, seagrass_agg[, c('num_seagrass', 'monthyear')], by = "monthyear")
features_agg <- left_join(features_agg, temp_agg[, c('monthyear', 'water_temp')], by="monthyear")

# add annual water temp
features_agg <- features_agg %>%
  group_by(year(monthyear)) %>%
  mutate(annual_water_temp = mean(water_temp))

# add seasonal water temp
features_agg <- features_agg %>%
  group_by(year(monthyear), Season) %>%
  mutate(seasonal_water_temp = mean(water_temp, na.rm = TRUE))

# rename year
features_agg <- features_agg %>%
  rename(year = 'year(monthyear)')

# select relevant columns and arrange rows by monthyear
features_agg <- features_agg %>%
  select(monthyear, Season, year, percent_coral_cover, num_fish, num_seagrass, water_temp, seasonal_water_temp, annual_water_temp) %>%
  arrange(monthyear)
```

```{r}
# save
setwd("cleaned_data")
write_csv(features_agg, "data_monthly_MM.csv")

# clear environment
rm(list = ls())
```
