==> 6_Shiny_app.Rmd <==
---
title: "RShiny App"
author: "Cheryn Aouaj, Lisa Bebell, Sun Kim, Manasi Mohan"
date: "12/13/2020"
output: html_document
runtime: shiny
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

We created an animation time-lapse RShiny app illustrating the changes in average water temperature at 9m depth collected at Moore Reef over time, along with the change in percentage of hard coral cover over time. As a reference, we included a stagnant line plot showing the change in global average temperature over time. We learned that there is an inverse relationship between the average temperature and hard coral cover. Throughout the years, as the water temperature of Moore Reef increased, hard coral cover decreased. This relationship is seen most dramatically in recent years, most notably around 2010. We can also see that the average global temperature change is increasing with each year. This information solidifies the need to take a proactive stance against climate change to save our planet and the species that inhabit our land and sea.

```{r eruptions, echo=FALSE}
library(shiny)
library(openxlsx)
library(stringr)
library(shinydashboardPlus)
library(shinydashboard)
library(lubridate)
library(ggplot2)
library(reshape2)

# load data
setwd("cleaned_data")
df <- read.xlsx("temperature_and_coral_cover.xlsx") 
climate <- read.csv("global.csv")
climate <- climate[, c("Year", "Average")]

# reformat date
df$date<-convertToDate(df$date)

# rename columns
colnames(df)[2]<-"Water"
colnames(df)[3]<-"Coral"

df_long <- melt(df, id="date")


ui <- dashboardPage(
    dashboardHeader(title = "Moore Reef Temperature and Hard Coral Cover Relationship Dashboard"),
    dashboardSidebar(disable = TRUE),
    dashboardBody(
        fluidRow(
            box(
                column(12,align="center",plotOutput("plot1", height = 350)),
                column(12,align="center",sliderInput("sliderA","",
                                                     min=min(as.Date(df$date)), max=max(as.Date(df$date)), step=30, value=as.Date("1997-11-21"),
                                                     animate = animationOptions(interval = 300, loop = TRUE)))),
            box(plotOutput("plot2"))
            
        )
    )
)


server <- function(input, output) {

    df_2<- reactive({
        df_long %>% dplyr::filter(date >= min(date) & date <= input$sliderA) })
    
    output$plot1 <- renderPlot({
        
        ggplot(df_2(),aes(x=date, y=value, colour=variable)) +
            geom_line(size=1.5)+
            scale_color_manual(values=c("#a0cf8d", "#89cff0"))+
            ylim(14,31) +
            ggtitle("Average Water Temperature and Mean Live Coral Cover Percentage")+
            theme(panel.background = element_blank(),
                  plot.title = element_text(size=12, face="bold",hjust = 0.5),
                  legend.position="top",
                  axis.ticks = element_blank(),
                  axis.title = element_blank(),
                  axis.text.x = element_text(face = "bold",size = 11 ,angle = 45),
                  axis.text.y = element_text(face = "bold",size = 12),
                  legend.background = element_rect(fill = "white"),legend.key = element_rect(fill = "white"),
                  legend.title = element_blank())
        
    })
    
    output$plot2 <- renderPlot({
        
        ggplot(climate,aes(x=Year, y=Average, group=1)) +
            ggtitle("Global Temperature Yearly Increase")+
            geom_line(size=1,color="#2565AE")+
            theme(panel.background = element_blank(),
                  plot.title = element_text(size=14, face="bold",hjust = 0.5),
                  axis.ticks = element_blank(),
                  axis.title = element_blank(),
                  axis.text.x = element_text(face = "bold",size = 11 ,angle = 45),
                  axis.text.y = element_text(face = "bold",size = 12),
                  legend.background = element_rect(fill = "white"),legend.key = element_rect(fill = "white"),
                  legend.title = element_blank())
        
    })
    
}

# Run the application 
shinyApp(ui, server, options = list(height = 650))
```
